requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
-f https://storage.googleapis.com/jax-releases/jax_releases.html
absl-py==1.0.0
asdfghjkl @ git+https://github.com/kazukiosawa/asdfghjkl.git@a94d4ea849d2873ed1b0fab8dc3e06b73bb1ea20
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bayesian-optimization==1.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
chex==0.1.3
click==8.0.4
cloudpickle==2.1.0
ConfigArgParse==1.5.3
ConfigSpace==0.4.21
cycler==0.11.0
Cython==0.29.28
decorator==5.1.0
Deprecated==1.2.13
distrax==0.1.2
dm-haiku==0.0.6
dm-tree==0.1.7
etils==0.6.0
filelock==3.6.0
flatbuffers==2.0
flax==0.5.0
fonttools==4.28.2
gast==0.5.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19
grpcio==1.42.0
h5py==3.6.0
higher==0.2.1
hpbandster==0.7.4
idna==3.3
immutabledict==2.2.1
importlib-metadata==4.8.2
importlib-resources==5.8.0
ipython==7.29.0
jax==0.3.14
jaxlib==0.3.14+cuda11.cudnn805
jedi==0.18.1
jmp==0.0.2
joblib==1.1.0
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kfac-jax @ git+https://github.com/deepmind/kfac-jax@8ca86829cc535d8775659471affc3ae2ab50a75a
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.0
matplotlib-inline==0.1.3
msgpack==1.0.3
netifaces==0.11.0
numexpr==2.8.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pandas==1.4.1
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
py-spy==0.3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
Pyro4==4.82
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ray==1.11.0
redis==4.1.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
serpent==1.40
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
tables==3.7.0
tabulate==0.8.9
tbparse==0.0.6
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.16.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tomli==1.2.2
toolz==0.11.2
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchviz==0.0.2
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.14.0
zipp==3.6.0
